+++ lib/validators/plugin_validator.rb
@@ -1,4 +1,8 @@
-class PluginValidator < ActiveModel::EachValidator  
+class PluginValidator < ActiveModel::EachValidator
+  def validate_each(record, attribute, value)
+    record.errors.add attribute, (options[:message] || validation_message) unless (allows_nil? && value.nil?) || is_valid?(value)
+  end
+
   protected
 
   def allows_nil?()
@@ -12,7 +16,13 @@ class PluginValidator < ActiveModel::EachValidator
     end
   end
 
-  def message_or(message)
-    return options[:message] || "message"
+  # Called to validate the value:
+  def is_valid?(value)
+    raise NotImplementedError, 'PluginValidator is not designed to be used directly'
+  end
+
+  # Returns the default validation message:
+  def validation_message
+    raise NotImplementedError, 'PluginValidator is not designed to be used directly'
   end
 end