+++ lib/validators/timezone_validator.rb
@@ -1,11 +1,11 @@
 class TimezoneValidator < PluginValidator
-  def validate_each(record, attribute, value)
-    record.errors.add attribute, message_or('is not a valid timezone') unless is_valid?(value)
-  end
-
   protected
 
+  def validation_message
+    'is not a valid time zone.'
+  end
+
   def is_valid?(value)
-    ActiveSupport::TimeZone.all.map(&:name).include?(value) || (allows_nil? && value.nil?)
+    ActiveSupport::TimeZone.all.map(&:name).include?(value)
   end
 end